chore: move the helper from script to macros
