+++ crates/fervid_transform/src/script/setup/macros.rs
@@ -1,7 +1,7 @@
 use fervid_core::{IntoIdent, VueImports};
 use swc_core::{
     common::DUMMY_SP,
-    ecma::ast::{ArrayLit, CallExpr, Callee, Expr, ExprOrSpread, Ident, ObjectLit, PropOrSpread},
+    ecma::ast::{ArrayLit, CallExpr, Callee, Decl, Expr, ExprOrSpread, Ident, Module, ModuleItem, ObjectLit, Pat, PropOrSpread, Stmt},
 };
 
 use crate::{
@@ -13,19 +13,152 @@ use crate::{
     script::{
         resolve_type::TypeResolveContext,
         setup::{
-            define_emits::process_define_emits,
-            define_model::process_define_model,
-            define_options::process_define_options,
-            define_props::{process_define_props, process_with_defaults},
-            define_slots::process_define_slots,
-            utils::unwrap_ts_node_expr,
+            define_emits::process_define_emits, define_model::process_define_model, define_options::process_define_options, define_props::{process_define_props, process_with_defaults}, define_props_destructure::collect_props_destructure, define_slots::process_define_slots, utils::unwrap_ts_node_expr
         },
     },
     structs::SfcExportedObjectHelper,
-    SetupBinding,
+    SetupBinding, TransformSfcContext,
 };
 
 use super::define_model::postprocess_models;
+
+pub struct CollectMacrosResult {
+    pub has_type_only_macros: bool,
+}
+
+pub fn collect_macros(
+    ctx: &mut TransformSfcContext,
+    module: &Module,
+    errors: &mut Vec<TransformError>,
+) -> CollectMacrosResult {
+    let mut result = CollectMacrosResult {
+        has_type_only_macros: false,
+    };
+
+    // Check `defineProps` inside `withDefaults`
+    // TODO Use a better approach
+    fn is_type_only_with_defaults(call_expr: &CallExpr) -> bool {
+        let Some(ExprOrSpread {
+            expr: define_props_expr,
+            spread: None,
+        }) = call_expr.args.first()
+        else {
+            return false;
+        };
+
+        let Expr::Call(CallExpr {
+            callee: Callee::Expr(callee_expr),
+            type_args,
+            ..
+        }) = define_props_expr.as_ref()
+        else {
+            return false;
+        };
+
+        let Expr::Ident(callee_expr_ident) = callee_expr.as_ref() else {
+            return false;
+        };
+
+        if callee_expr_ident.sym == *DEFINE_PROPS {
+            type_args.is_some()
+        } else {
+            false
+        }
+    }
+
+    for module_item in module.body.iter() {
+        let ModuleItem::Stmt(module_stmt) = module_item else {
+            continue;
+        };
+
+        match module_stmt {
+            // E.g. `let foo = defineProps()`
+            Stmt::Decl(Decl::Var(var_decl)) => {
+                if var_decl.declare {
+                    continue;
+                }
+
+                for var_decl_item in var_decl.decls.iter() {
+                    let Some(ref init) = var_decl_item.init else {
+                        continue;
+                    };
+
+                    let Expr::Call(ref call_expr) = init.as_ref() else {
+                        continue;
+                    };
+
+                    let Callee::Expr(ref callee_expr) = call_expr.callee else {
+                        continue;
+                    };
+
+                    let Expr::Ident(ref callee_ident) = callee_expr.as_ref() else {
+                        continue;
+                    };
+
+                    let is_define_props = callee_ident.sym == *DEFINE_PROPS;
+                    let is_define_emits = callee_ident.sym == *DEFINE_EMITS;
+                    let is_with_defaults = callee_ident.sym == *WITH_DEFAULTS;
+                    if !is_define_props && !is_define_emits && !is_with_defaults {
+                        continue;
+                    }
+
+                    // Type-only marker
+                    let has_type_args = if is_define_props || is_define_emits {
+                        call_expr.type_args.is_some()
+                    } else {
+                        is_type_only_with_defaults(call_expr)
+                    };
+
+                    result.has_type_only_macros |= has_type_args;
+
+                    // No other logic for `defineEmits`
+                    if is_define_emits {
+                        continue;
+                    }
+
+                    // Props destructure supports only object patterns
+                    if let Pat::Object(ref obj_pat) = var_decl_item.name {
+                        collect_props_destructure(ctx, obj_pat, errors)
+                    }
+                }
+            }
+
+            Stmt::Expr(expr_stmt) => {
+                let Expr::Call(ref call_expr) = expr_stmt.expr.as_ref() else {
+                    continue;
+                };
+
+                let Callee::Expr(ref callee_expr) = call_expr.callee else {
+                    continue;
+                };
+
+                let Expr::Ident(ref callee_ident) = callee_expr.as_ref() else {
+                    continue;
+                };
+
+                let is_define_props = callee_ident.sym == *DEFINE_PROPS;
+                let is_define_emits = callee_ident.sym == *DEFINE_EMITS;
+                let is_with_defaults = callee_ident.sym == *WITH_DEFAULTS;
+                if !is_define_props && !is_define_emits && !is_with_defaults {
+                    continue;
+                }
+
+                // Type-only marker
+                let has_type_args = if is_define_props || is_define_emits {
+                    call_expr.type_args.is_some()
+                } else {
+                    is_type_only_with_defaults(call_expr)
+                };
+
+                result.has_type_only_macros |= has_type_args;
+            }
+
+            _ => continue,
+        };
+    }
+
+    result
+}
 
 pub enum TransformMacroResult {
     NotAMacro,
